% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta.sim.R
\name{Beta.sim}
\alias{Beta.sim}
\title{Beta Simulator}
\usage{
Beta.sim(n=1337,a=2.5,b=5.5,initVal=0.5)
}
\arguments{
\item{n}{is a natural number that chooses the number of iterations for the simulation.}

\item{a}{is a non-negative number that chooses the first parameter of the Beta distribution.}

\item{b}{is a non-negative number that chooses the second parameter of the Beta distribution.}

\item{initVal}{is a single number between 0 and 1 that chooses the starting value for the simulation.}
}
\value{
Returns a vector of numbers to be Beta(a,b) distribution, using the Metropolis-Hastings sampler.
}
\description{
This function takes 0 to 4 arguments to use the Metropolis-Hastings sampler to generate Beta distributions.
If no arguments are used, 1337 iterations are used to generate a Beta(2.5,5.5) distribution using uniform(0,1) distribution.
The function also always prints the rejection rate.
}
\examples{
Beta.sim()
Beta.sim(n=200)
}
\author{
Nguyen Khanh Le Ho & Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
\email{ngho14@student.sdu.dk} \cr
}
\keyword{Beta}
\keyword{distribution}
\keyword{hastings}
\keyword{iteration}
\keyword{metropolis}
\keyword{sampler}
\keyword{simulation}
\keyword{uniform}

