% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.correlation.R
\name{bootstrap.correlation}
\alias{bootstrap.correlation}
\title{Bootstrap Estimation of Correlation}
\usage{
bootstrap.correlation(n=200, x , y, plot = FALSE)
}
\arguments{
\item{n}{is a positive integer that specifies how many bootstrap replicates should be used for estimation.}

\item{x}{is a vector of numeric values to be used with y to estimate correlation and bootstrap estimates.}

\item{y}{is a vector of numeric values to be used with x to estimate correlation and bootstrap estimates. y has to be of the same length as x.}

\item{plot}{is a boolean value that specifies whether a plot should be created or not}
}
\value{
Returns a list containing: the estimated correlation of the two given vectors, the bootstrap estimated correlation of the data, as well as the standard error, bias and the 95 percent confidence interval.
}
\description{
This function takes as arguments the number of bootstrap replicates \code{n},
\code{x} and \code{y} which are two numerical vectors / data, then bootstrap from \code{x ,y}, and calculate the correlation of these data for each resampling n times.
If \code{plot} is chosen to be TRUE, then the function plots the histogram, as well as the 95% confidence interval for the
estimates.
}
\examples{
bootstrap.correlation(n=2000,cars$speed,cars$dist, plot=TRUE )
}
\author{
Nguyen Khanh Le Ho & Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
\email{ngho14@student.sdu.dk} \cr
}
\keyword{bias}
\keyword{bootstrap}
\keyword{confidence}
\keyword{correlation}
\keyword{error}
\keyword{estimation}
\keyword{interval}
\keyword{standard}

