% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.needle.R
\name{buffon.needle}
\alias{buffon.needle}
\title{Buffon's needle}
\usage{
buffon.needle(n=10, l=1, d=1)
}
\arguments{
\item{n}{is a positive integer value that specifies how many needles that are tossed in the simulation.}

\item{l}{is a positive number that specifies the length of the needles tossed in the simulation.}

\item{d}{is a positive number that specifies the distance between the stripes on the floor in the simulation.}
}
\value{
returns pi estimated by simulating Buffon's experiment.
}
\description{
This function takes 3 arguments to simulate Buffon's experiment to estimate pi.
}
\examples{
buffon.needle(n=3408, l=2.5,d=3)
}
\author{
Nguyen Khanh Le Ho & Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
\email{ngho14@student.sdu.dk} \cr
}
\keyword{buffon}
\keyword{estimation}
\keyword{needle}
\keyword{pi}
\keyword{simulation}

