% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.probability.R
\name{chi.probability}
\alias{chi.probability}
\title{Calculation the Chi Squared probability.}
\usage{
chi.probability(x, df = 5, n = 1000)
}
\arguments{
\item{x}{specifies the value that should be evaluated in the Chi Square distribution.}

\item{df}{specifies degrees of freedom in the Chi Square distribution.}

\item{n}{specifies number of random numbers for evaluating the distribution.}
}
\value{
returns the probability that the \code{Chi} is greater than \code{x} (area under the distribution which is greater than \code{x}).
}
\description{
\code{chi.probability} computes the probability that the \code{Chi} of a given degree of freedom is greater than a specified value \code{x}.
If \code{x} is a vector, the function will return a vector of probabilities of the components of \code{x}.
}
\examples{
chi.probability(x= 5, df = 5, n = 1000)
}
\author{
Nguyen Khanh Le Ho & Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
\email{ngho14@student.sdu.dk} \cr
}
\keyword{Carlo}
\keyword{Estimation}
\keyword{Monte}
\keyword{chi}
\keyword{distribution}
\keyword{simulation}

