% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.plot.R
\name{dens.plot}
\alias{dens.plot}
\title{Density Plotter}
\usage{
dens.plot(x,n=500,method="naive",from,to)
}
\arguments{
\item{x}{is data given as a vector of numbers, to estimate density from.}

\item{n}{is a positive integer that specifies the number of points to be used for plotting the density function. Defaults to 500.}

\item{method}{is an argument that is either "naive" or "kernel", deciding the method of estimating the density. Defaults to "naive".}

\item{from}{is a number that specifies where the plot should start. Defaults to min(x)-sd(x)/3}

\item{to}{is a number that specifies where the plot should end. Defaults to max(x)+sd(x)/3}
}
\value{
Returns a density plot over the input data x, with information calculated depending on other inputs, or lack thereof.
}
\description{
This function takes between 1 and 5 arguments to plot the density of data, using the dens.estimator function.
}
\examples{
dens.plot(cars$speed)
dens.plot(cars$speed,method="kernel")
dens.plot(cars$speed,n=512,method="naive", from=10, to=20)
}
\author{
Nguyen Khanh Le Ho & Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
\email{ngho14@student.sdu.dk} \cr
}
\keyword{bandwidth}
\keyword{data}
\keyword{density}
\keyword{estimate}
\keyword{gaussian}
\keyword{kernel}
\keyword{naive}
\keyword{plot}
\keyword{silverman}
\keyword{struges}
\keyword{sturges}

