% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.model.R
\name{linear.model}
\alias{linear.model}
\title{Linear Model summarizer}
\usage{
linear.model(formula)
}
\arguments{
\item{formula}{is an argument of the type 'language' that apply the syntax: dependent~predictor1+predictor2+...+predictorN}
}
\value{
Returns a list containing: a vector of residuals, a single value for Residual Standard Error, single value for R-squared,
single value for adjusted R-squared, single value for F-statistic, single value for degrees of freedom, single value for p-value,
And a data.frame of coefficients(Estimates, standard error, t-value, PR(>|t|))
}
\description{
This function takes as argument a formula with a certain syntax, dependent~predictor1+predictor2+...+predictorN,
And returns a list of information including residuals and coefficients(Estimate, standard error, t-value, Pr(>|t|)) in data.frames,
as well as the residual standard error, degrees of freedom, Multiple R-squared, Adjusted R-squared, F-statistic and the p-value.
}
\examples{
linear.model(cars$speed~cars$dist)
linear.model(cars$speed~cars$dist+I(cars$dist^2))
linear.model(cars$speed~cars$dist+I(cars$dist^2)+I(cars$dist*cars$dist*cars$dist))
}
\author{
Nguyen Khanh Le Ho & Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
\email{ngho14@student.sdu.dk} \cr
}
\keyword{degrees}
\keyword{df}
\keyword{error}
\keyword{estimate}
\keyword{freedom}
\keyword{linear}
\keyword{lm}
\keyword{model}
\keyword{p}
\keyword{residual}
\keyword{standard}
\keyword{summary}
\keyword{t}
\keyword{value}

